Move fixtures to tests
